Cast NSNumber to Float

+++ ios/Classes/BarometerStreamHandler.swift
@@ -37,7 +37,7 @@ public class BarometerStreamHandler : NSObject, FlutterStreamHandler {
             guard error == nil else { return }
             guard let altitudeData = data else { return }
             // Convert from kilopascals to hPa
-            let dataArray = [ altitudeData.pressure * 10 ]
+            let dataArray = [ altitudeData.pressure.floatValue * 10 ]
             SwiftFlutterSensorsPlugin.notify(sensorId: BarometerStreamHandler.SENSOR_ID, sensorData: dataArray, eventSink: eventSink)
         })
     }
